+++ src/main/java/ua/yandex/shad/stream/IntStream.java
@@ -13,15 +13,15 @@ import ua.yandex.shad.function.IntBinaryOperator;
  */
 public interface IntStream {
 
-    Double average();
+    double average();
 
-    Integer max();
+    int max();
 
-    Integer min();
+    int min();
     
     IntStream flatMap(IntToIntStreamFunction func);
 
-    long count();
+    int count();
 
     IntStream filter(IntPredicate predicate);
 
@@ -31,7 +31,7 @@ public interface IntStream {
 
     int reduce(int identity, IntBinaryOperator op);
 
-    Integer sum();
+    int sum();
 
     int[] toArray();
 }
